% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMNIPath}
\alias{getMNIPath}
\title{Get MNI Path of File}
\usage{
getMNIPath(what = c("T1", "Brain", "Brain_Mask"), res = c("1mm",
  "2mm"))
}
\arguments{
\item{what}{Type of image to return}

\item{res}{Resolution of the image, in millimeters}
}
\value{
Character filename of path
}
\description{
Returns the path of a MNI file to the user.
}
\examples{
getMNIPath()
}
