% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMNISegPath}
\alias{getMNISegPath}
\title{Get MNI Path of Segmentation File}
\usage{
getMNISegPath(alg = "FAST", res = c("1mm", "2mm"))
}
\arguments{
\item{alg}{Algorithm to do the segmentation}

\item{res}{Resolution of the image, in millimeters}
}
\value{
Character filename of path
}
\description{
Returns the path of a MNI file of a segmentation to the user.
}
\examples{
getMNIPath()
}

